city_project/requirements/base.txt
beautifulsoup4==4.7.1
Django==2.1.4
django-appconf==1.0.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-leaflet==0.24.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
googlemaps==3.0.2
gunicorn==19.9.0
Pillow==5.4.1
postgis==1.0.4
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint-django==2.0.5
pytest-mocha==0.2.1
python-google-places==1.4.1
pytz==2018.9
requests==2.21.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
typed-ast==1.3.0
python-decouple==3.1
